Support macOS

+++ src/valhalla_associate_segments.cc
@@ -839,7 +839,7 @@ void edge_association::add_tile(const std::string &file_name) {
 
   //do the matching of the segments in this osmlr tile
   std::cout.precision(16);
-  size_t entry_id = 0;
+  uint64_t entry_id = 0;
   for (auto &entry : tile.entries()) {
     if (!entry.has_marker()) {
       assert(entry.has_segment());
